playerscore is double

+++ app/src/main/java/app/meeplebook/core/plays/domain/CreatePlayerCommand.kt
@@ -20,6 +20,6 @@ data class CreatePlayerCommand(
     val name: String,
     val startPosition: Int?,
     val color: String?,
-    val score: Int?,
+    val score: Double?,
     val win: Boolean
 )
